Significant reworking of object polymorph.
The changes are as follows. Some could be split out, but it seemed easier to do
a monolithic commit since this is all roughly the same area of code.

- Amulets of unchanging no longer can be polymorphed or be the result of
  polymorph.
- Potions of polymorph, when dipped, now perform regular alchemy rather than
  being polymorphed themselves.
- When a wielded or worn item is polymorphed, it remains wielded or worn if it
  fits in the same or similar (armor/shirt/cloak) slot and can be.

Known bugs:
- Twoweaponing isn't handled well (this is a bug everywhere, though...)

+++ libnethack/src/invent.c
@@ -731,7 +731,7 @@ putting_on(const char *action)
 static enum obj_use_status
 object_selection_checks(struct obj *otmp, const char *word)
 {
-    long dummymask;
+    int dummymask;
     int otyp = otmp->otyp;
 
     /* ugly check: remove inappropriate things */
